+++ tests/test_html.py
@@ -15,7 +15,7 @@ config = {
             'element': 'li',
             'wrapper': ['ul', {'class': 'public-DraftStyleDefault-ul'}]
         },
-        'unstyled': {'element': 'div'}
+        'unstyled': {'element': 'p'}
     },
     'style_map': {
         'ITALIC': {'element': 'em'},
@@ -193,15 +193,15 @@ class TestHTML(unittest.TestCase):
             ]
         })), str([
             Command('start_text', 0),
-            Command('stop_text', 19),
             Command('start_inline_style', 0, 'ITALIC'),
+            Command('start_entity', 0, 1),
             Command('stop_inline_style', 4, 'ITALIC'),
+            Command('stop_entity', 4, 1),
+            Command('start_entity', 5, 0),
             Command('start_inline_style', 9, 'BOLD'),
             Command('stop_inline_style', 12, 'BOLD'),
-            Command('start_entity', 5, 0),
             Command('stop_entity', 14, 0),
-            Command('start_entity', 0, 1),
-            Command('stop_entity', 4, 1),
+            Command('stop_text', 19),
         ]))
 
     def test_build_command_groups_empty(self):
@@ -292,6 +292,16 @@ class TestHTML(unittest.TestCase):
                 },
             ]
         }), '<h1>Header</h1>')
+
+    def test_render_block_defaults(self):
+        self.assertEqual(self.exporter.render({
+            'entityMap': {},
+            'blocks': [
+                {
+                    'text': 'Paragraph',
+                },
+            ]
+        }), '<p>Paragraph</p>')
 
     def test_render_empty(self):
         self.assertEqual(self.exporter.render({