+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## Unreleased
 
+## [v2.1.6](https://github.com/springload/draftjs_exporter/releases/tag/v2.1.6)
+
+### Changed
+
+* Assume same block defaults as Draft.js would when attributes are missing: depth = 0, type = unstyled, no entities, no styles ([#110](https://github.com/springload/draftjs_exporter/pull/110), thanks to [@tpict](https://github.com/tpict)).
+* Minor performance improvements for text-only blocks ([#112](https://github.com/springload/draftjs_exporter/pull/112)).
+
 ## [v2.1.5](https://github.com/springload/draftjs_exporter/releases/tag/v2.1.5)
 
 ### Changed
@@ -400,7 +407,7 @@ KEYBOARD = 'kbd'
 
 ### Added
 
-* Add support for decorators thanks to @su27 (#16, #17).
+* Add support for decorators thanks to [@su27](https://github.com/su27) (#16, #17).
 * Add support for configurable decorators and entities.
 * Add support for decorators and entities in function form.
 
@@ -416,7 +423,7 @@ KEYBOARD = 'kbd'
 
 ### Added
 
-* Add profiling tooling thanks to @su27 (#31).
+* Add profiling tooling thanks to [@su27](https://github.com/su27) (#31).
 * Add more common entity types in constants (#34).
 
 ### Fixed